safari-600.7-branch/LayoutTests/platform/ios-sim/accessibility/popup-button-value-label-expected.txt

  
This tests that a popup button will return the selected item through its value attribute instead of its label attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS pass.iphoneValue is 'AXValue: item 1'
PASS pass.iphoneLabel is 'AXLabel: '
PASS pass.iphoneValue is 'AXValue: item 2'
PASS pass.iphoneLabel is 'AXLabel: Popup'
PASS successfullyParsed is true

TEST COMPLETE
